[Bug]위시리스트 추가 상세정보페이지 반영 x, 콘솔제거, 페이지네이션...
